Improve UI layout and functionality using Bootstrap and vanilla JS.
Using Bootstrap and some custom CSS, we designed the frontend layout
of the application. Using vanilla JS, we added a few event listeners
to our application——mainly to the task action icons. A couple of them
will submit the parent form that the icon sits inside when clicked.
Another simply adds an existing Bootstrap class to the icon to change
the color of the icon.