Work on application layout and UI
When hovering over a task item, the item is not highlighted with
a grey color. Back buttons replace the previous hard coded links
and takes our user back to the last page based on the browser's
history. The 'Mark as Completed' and 'Delete' buttons in the task
show page are now functional and working.